Consolidating work, deleting old SearchFacet component
