+++ src/assets/scss/default.scss
@@ -515,15 +515,15 @@ section#main {
 					}
 				}
 
-				.filter-field {
-					margin-bottom: 0.5rem;
-				}
-
 				ul {
 					padding-left:1.5rem;
 					li {
 						list-style-type: none;
 					}
+				}
+
+				.filter-field {
+					margin-bottom: 0.5rem;
 				}
 
 				.count {
