+++ src/explorer/index.js
@@ -1,4 +1,3 @@
 
 export { default as SearchFacet } from './SearchFacet';
-export { default as SearchFacet2 } from './SearchFacet2';
 export { default as ResultsList } from './ResultsList';
